Commit testes unitários e ajustes.

+++ src/main/java/com/junior/personcadastre/service/PersonPersistenceService.java
@@ -1,7 +1,9 @@
 package com.junior.personcadastre.service;
 
+import com.junior.personcadastre.api.mapper.PersonMapper;
 import com.junior.personcadastre.domain.Person;
 import com.junior.personcadastre.exception.BadRequestException;
+import com.junior.personcadastre.exception.NotFoundException;
 import com.junior.personcadastre.repository.PersonRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -9,6 +11,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.Optional;
 
 @Service
 public class PersonPersistenceService {
@@ -18,26 +22,26 @@ public class PersonPersistenceService {
     @Autowired
     private PersonRepository personRepository;
 
-    @Transactional(propagation = Propagation.REQUIRED)
-    public Person saveOrUpdatePerson(Person personSave) throws Exception {
-        Person person;
-        try {
-            person = personRepository.save(personSave);
-        } catch (Exception e) {
-            LOG.error(e.getMessage());
-            throw new BadRequestException("Ocorreu um erro ao salvar a pessoa ID: " + personSave.getPersonId());
-        }
+    @Autowired
+    private PersonSearchService personSearchService;
 
-        return person;
+    @Transactional(propagation = Propagation.REQUIRED)
+    public Person saveOrUpdatePerson(Person personSave) {
+        return personRepository.save(personSave);
     }
 
     @Transactional(propagation = Propagation.REQUIRED)
-    public void removePerson(int personId) throws Exception {
-        try {
+    public String removePerson(int personId) throws Exception {
+
+        Optional<Person> personList = personRepository.findById(personId);
+
+        if (!personList.isPresent()) {
+            throw new BadRequestException("Error deleting person ID: " + personId);
+        } else {
             personRepository.deleteById(personId);
-        } catch (Exception e) {
-            LOG.error(e.getMessage());
-            throw new BadRequestException("Ocorreu um erro ao excluir a pessoa ID: " + personId);
+
+            String message = "Person ID: " + personId + " removed with success!.";
+            return message;
         }
     }
 }
